refactor: reorganize some firebase files
