+++ app/login/layout.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 import type { ReactNode } from "react";
-import { getAuthenticatedUser } from "../../lib/Firebase/serverAuth";
+import { getAuthenticatedUser } from "../../lib/Firebase/server/auth";
 import { ROUTES } from "utils/routes";
 
 export default async function LoginLayout({
