+++ app/page.tsx
@@ -1,5 +1,5 @@
 import { EditionYear } from "@/components/EditionYear";
-import { getEditions } from "../lib/Firebase/firebaseServersideAPIs";
+import { getEditions } from "../lib/Firebase/server/api";
 
 // Never revalidated, similar to getStaticProps
 export const revalidate = false;
