+++ lib/Firebase/client/firebaseAuthServiceWorker.ts
