+++ lib/Firebase/client/hooks.ts
@@ -18,7 +18,7 @@ import {
 } from "firebase/firestore";
 import { useState } from "react";
 import { useCollectionData } from "react-firebase-hooks/firestore";
-import { IArticle, ISettings } from "../types";
+import { IArticle, ISettings } from "../../types";
 import { db } from "./firebase";
 
 const articleConverter = {
