+++ lib/Firebase/server/auth.ts
